Move 2 fixtures to conftest.py
